+++ NodeJs/src/Views/RanchoApp.js
@@ -1,37 +1,61 @@
 import React, { useState, useEffect } from 'react';
 import RanchoService from '../Controllers/RanchoService';
 import { Link } from 'react-router-dom';
+
 export default function RanchoApp() {
     const [rancho, setRancho] = useState([]);
+
+    const capitalize = (str) => {
+        return str.replace(/\w\S*/g, (txt) => txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase());
+    };
+
     const listarCarga = () => {
         RanchoService.findAll().then((response) => {
-            setRancho(response.data);
-            console.log(response.data);
+            const formatoData = response.data.map(item => ({
+                ...item,
+                nombre_Rancho: capitalize(item.nombre_Rancho),
+                ubicacion_Rancho: capitalize(item.ubicacion_Rancho),
+                extension_Rancho: capitalize(item.extension_Rancho)
+            }));
+            setRancho(formatoData);
+            console.log(formatoData);
         }).catch((error) => {
             console.log(error);
         });
-    }
+    };
+
     useEffect(() => {
         listarCarga();
     }, []);
 
     useEffect(() => {
         RanchoService.findAll().then((response) => {
-            setRancho(response.data);
-            console.log(response.data);
+            const formatoData = response.data.map(item => ({
+                ...item,
+                nombre_Rancho: capitalize(item.nombre_Rancho),
+                ubicacion_Rancho: capitalize(item.ubicacion_Rancho),
+                extension_Rancho: capitalize(item.extension_Rancho)
+            }));
+            setRancho(formatoData);
+            console.log(formatoData);
         }).catch((error) => {
             console.log(error);
         });
-
     }, []);
+
     const deleteRancho = (id) => {
-        RanchoService.delete(id).then((response) => {
-            listarCarga();
-            console.log(response.data);
-        }).catch((error) => {
-            console.log(error);
-        });
+        const confirmarDelete = window.confirm("¿Estás seguro de que deseas eliminar este registro?");
+        if (confirmarDelete) {
+            RanchoService.delete(id).then((response) => {
+                listarCarga();
+                console.log(response.data);
+            }).catch((error) => {
+                console.log(error);
+            });
+        }
+
     }
+
     return (
         <div>
             <footer className="tittleFrm">Rancho</footer>
